+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.2.0 (Next)
 
+* Sending `send_message` with nil or empty text will yield `Nothing to see here.` with a GIF instead of `no_text` - [@dblock](https://github.com/dblock).
 * Added support for operators with `operator [name]` - [@dblock](https://github.com/dblock).
 * Added support for custom commands with `command [name]` - [@dblock](https://github.com/dblock).
 * Your contribution here.